feat: hydrate
